+++ cases/testcases/implicit_dependencies/explicit_require/green1.clj
@@ -1,0 +1,6 @@
+(ns testcases.implicit-dependencies.explicit-require.green1
+  "Exercises an explicit `require`, with aliasing")
+
+(require '[clojure.string :as string])
+
+string/blank?