+++ cases/testcases/implicit_dependencies/explicit_require/green2.clj
@@ -1,0 +1,6 @@
+(ns testcases.implicit-dependencies.explicit-require.green2
+  "Exercises an explicit `require`, without aliasing")
+
+(require 'clojure.string)
+
+clojure.string/blank?