+++ cases/testcases/implicit_dependencies/explicit_require/green4.clj
@@ -1,0 +1,6 @@
+(ns testcases.implicit-dependencies.explicit-require.green4
+  "Exercises prefix notation, without aliases")
+
+(require '[clojure [string]])
+
+clojure.string/blank?