+++ cases/testcases/implicit_dependencies/explicit_require/green5.clj
@@ -1,0 +1,6 @@
+(ns testcases.implicit-dependencies.explicit-require.green5
+  "Exercises prefix notation, with aliases")
+
+(require '[clojure [string :as s]])
+
+s/blank?