+++ cases/testcases/implicit_dependencies/explicit_require/red.clj
@@ -1,0 +1,3 @@
+(ns testcases.implicit-dependencies.explicit-require.red)
+
+clojure.string/blank?